pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drifts-and-complexity"
version = "0.1.0"
description = "Analysis of concept drift and process complexity in event logs"
authors = [
    {name = "Lennart Ebert", email = "lennart.ebert@hu-berlin.de"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=1.5.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.2.0",
    "networkx>=3.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "tqdm>=4.64.0",
    "pytz>=2022.7.0",
    "pyyaml",
    "lxml",
    "graphviz",
    "python-graphviz",
    "pm4py>=2.4.1",
    "pulp>=3.2.1",
    "Wasserstein>=1.1.0",
    "strsimpy>=0.2.1",
    "Levenshtein>=0.21.0",
    "BitVector",
    "lempel_ziv_complexity",
    "python-Levenshtein",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
    "pre-commit>=3.0.0",
    "types-PyYAML",
    "types-requests",
    "types-setuptools",
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["utils*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
show_column_numbers = true
show_error_context = true
pretty = true
color_output = true
error_summary = true

[[tool.mypy.overrides]]
module = [
    "pm4py.*",
    "Levenshtein.*",
    "lxml.*",
    "graphviz.*",
    "matplotlib.*",
    "seaborn.*",
    "scipy.*",
    "sklearn.*",
    "networkx.*",
    "pandas.*",
    "numpy.*",
    "pulp.*",
    "Wasserstein.*",
    "strsimpy.*",
    "BitVector.*",
    "lempel_ziv_complexity.*",
    "python-Levenshtein.*",
]
ignore_missing_imports = true

[tool.mypy-overrides]
"tests.*" = {disallow_untyped_defs = false, disallow_incomplete_defs = false}
"plugins.*" = {ignore_errors = true}
"*.ipynb" = {ignore_errors = true}
"assess_datasets.py" = {disallow_untyped_defs = false}
"combine_results.py" = {disallow_untyped_defs = false}
"run_bias_problematization.py" = {disallow_untyped_defs = false}
"run_bias_study.py" = {disallow_untyped_defs = false}
"test.py" = {disallow_untyped_defs = false}

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["utils"]
known_third_party = ["pm4py", "pandas", "numpy", "matplotlib", "scipy", "sklearn"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--cov=utils",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
